Fixes drops

+++ Dungeons and Dragons/Assets/Scripts/PlayerFolder/PlayerMovement.cs
@@ -42,8 +42,8 @@ public class PlayerMovement : MonoBehaviour
         healthBar.SetSize(1.21f);
     }
 
-    public void increaseSpeed(float amount){
-        moveSpeed += amount;
+    public void increaseSpeed(double amount){
+        moveSpeed += (float)amount;
     }
